wetube save 07.26 14:57

+++ src/controller/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/video";
 
 export const home = async (req,res) => {
@@ -8,7 +9,7 @@ export const home = async (req,res) => {
 };
 export const watch = async (req,res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(video === null){
         return res.status(404).render("404",{ pageTitle : "Video Not Found",});
     };
@@ -16,19 +17,27 @@ export const watch = async (req,res) => {
 };
 export const getEdit = async (req,res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404",{ pageTitle : "Video Not Found",});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit" , {pageTitle :`Editing : ${video.title}`,video});
 };
 
 export const postEdit = async (req,res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const { title , description , hashtags} = req.body;
     const video = await Video.exists({_id : id});
     if (!video) {
         return res.render("404",{ pageTitle : "Video Not Found",});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -43,16 +52,20 @@ export const getUpload = (req,res) => {
 };
 
 export const postUpload = async (req,res) => {
-    const file = req.file;
+    const {user: {_id},} = req.session;
+    const {path: fileUrl} = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
-            fileUrl:file.path,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
-    
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/")
     } catch(error) {
         return res.status(400).render("upload",{
@@ -64,6 +77,14 @@ export const postUpload = async (req,res) => {
 
 export const deleteVideo = async (req,res)=> {
     const {id} = req.params;
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.render("404",{ pageTitle : "Video Not Found",});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
